vibesecurity/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.9.3
aiosqlite==0.20.0
slowapi==0.1.9
